"#382 compatible variable names"

+++ necst/procedures/measurement/knife_edge.py
@@ -1,5 +1,4 @@
-import time
-
+import numpy as np
 from .measurement_base import Measurement
 
 
@@ -9,7 +8,7 @@ class Knife_edge(Measurement):
     def run(
         self,
         id: list,
-        positions=np.linspace(4750, 19700, 300),
+        positions=np.linspace(4750, 19700, 300, dtype=int),
     ) -> None:
         for position in positions:
             self.com.chopper(cmd=position, wait_oc=True)
